ensure np.ceil output is a positive integer

+++ hetgpy/hetGP.py
@@ -1213,7 +1213,7 @@ class hetGP:
                 idx = idx + len(init['Delta'])
                 # reconcile Deltas with R implementation
                 if trace > 1:
-                    for ii in range(np.ceil(len(mle_par['Delta'])/5)):
+                    for ii in range(np.ceil(len(mle_par['Delta'])/5).astype(np.uint64)):
                         i_tmp = np.arange(1 + 5*(ii-1),min(5*ii, len(mle_par['Delta'])))
                         if logN: print("Delta |", mle_par['Delta'][i_tmp], " | ", np.max(np.log(eps * mult[i_tmp]), init['Delta'][i_tmp] - np.log(1000)), " | ", init['Delta'][i_tmp] + np.log(100), "\n")
                         if not logN: print("Delta |", mle_par['Delta'][i_tmp], " | ", np.max(mult[i_tmp] * eps, init['Delta'][i_tmp] / 1000), " | ", init['Delta'][i_tmp] * 100, "\n")
